Integrate visual editor into MainUI
